\name{pmc_identify}
\alias{pmc_identify}
\title{Learn about the PubMed Central OAI-PMH service.}
\usage{
  pmc_identify(transform = TRUE,
    url = "http://www.pubmedcentral.gov/oai/oai.cgi")
}
\arguments{
  \item{transform}{transform metadata to list (TRUE/FALSE)}

  \item{url}{the base url for DataCite (leave to default)}
}
\description{
  Learn about the PubMed Central OAI-PMH service.
}
\examples{
\dontrun{
pmc_identify()
}
}

